pyproject.toml
[project]
name = "hub-and-spoke"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "igraph>=1.0.0",
    "matplotlib>=3.9.4",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "polars>=1.36.1",
    "pyarrow>=21.0.0",
]
